+++ core/src/main/scala/anorm/ToStatement.scala
@@ -15,33 +15,12 @@ import java.math.{ BigDecimal => JBigDec, BigInteger }
 
 import java.sql.{ PreparedStatement, Timestamp }
 
-/** Sets value as statement parameter. */
-@annotation.implicitNotFound("Cannot set value of type ${A} as parameter on statement: `anorm.ToStatement[${A}] required`")
-trait ToStatement[A] { self =>
+private[anorm] trait ToStatementBase[A] { self =>
 
   /**
    * Sets value |v| on statement |s| at specified |index|.
    */
   def set(s: PreparedStatement, index: Int, v: A): Unit
-
-  /**
-    * {{{
-    * import anorm.ToStatement
-    * 
-    * sealed trait MyEnum
-    * case object Foo extends MyEnum
-    * case object Bar extends MyEnum
-    * 
-    * implicit val to: ToStatement[MyEnum] = 
-    *   ToStatement.of[Int].contramap[MyEnum] {
-    *     case Foo => 1
-    *     case Bar => 2
-    *   }
-    * }}}
-    */
-  final def contramap[B](f: B => A): ToStatement[B] =
-    ToStatement[B] { (s, i, b) => self.set(s, i, f(b)) }
-
 }
 
 /*
@@ -773,18 +752,8 @@ sealed trait ToStatementPriority1 extends ToStatementPriority0 {
 /**
  * Provided conversions to set statement parameter.
  */
-object ToStatement extends ToStatementPriority1
+private[anorm] class ToStatementConversions extends ToStatementPriority1
     with JodaToStatement with JavaTimeToStatement {
-
-  private class FunctionalToStatement[T](
-    f: (PreparedStatement, Int, T) => Unit
-  ) extends ToStatement[T] {
-    def set(s: PreparedStatement, index: Int, v: T) { f(s, index, v) }
-  }
-
-  /** Functional factory */
-  def apply[T](set: (PreparedStatement, Int, T) => Unit): ToStatement[T] =
-    new FunctionalToStatement[T](set)
 
   /**
     * Resolves `ToStatement` instance for the given type.
